Add missing include for gcc and fix a gcc warning.

+++ src/XrdOss/XrdOss.hh
@@ -14,6 +14,7 @@
 
 #include <dirent.h>
 #include <errno.h>
+#include <strings.h>
 #include <sys/stat.h>
 #include <sys/types.h>
 
@@ -138,7 +139,7 @@ virtual int     StatVS(XrdOssVSInfo *sP, const char *sname=0, int updt=0)
                       {return -ENOTSUP;}
 
 virtual int     Lfn2Pfn(const char *Path, char *buff, int blen)
-                       {if (strlen(Path) >= blen) return -ENAMETOOLONG;
+                       {if ((int)strlen(Path) >= blen) return -ENAMETOOLONG;
                         strcpy(buff, Path); return 0;
                        }
